pedritovm

+++ MissTortasRework/MissTortas.Engine/DTO/Products/CreateSaleProductDTO.cs
@@ -2,8 +2,8 @@
 {
     public class CreateSaleProductDTO
     {
-        public float SalePrice { get; set; }
-        public long SaleQuantity { get; set; }
+        public double SalePrice { get; set; }
+        public double SaleQuantity { get; set; }
         public long StockProductId { get; set; }
         public string SaleDescription { get; set; } = string.Empty;
     }
