+++ MissTortasRework/MissTortas.Engine/Validators/Products/CreateSaleProductDTOValidator.cs
@@ -7,8 +7,8 @@ namespace MissTortas.Engine.Validators.Products
     {
         public CreateSaleProductDTOValidator()
         {
-            RuleFor(dto => dto.SaleQuantity).NotEmpty().InclusiveBetween(1, long.MaxValue - 1);
-            RuleFor(dto => dto.SalePrice).NotEmpty().InclusiveBetween(1, float.MaxValue - 1);
+            RuleFor(dto => dto.SaleQuantity).NotEmpty().InclusiveBetween(1, double.MaxValue - 1);
+            RuleFor(dto => dto.SalePrice).NotEmpty().InclusiveBetween(1, double.MaxValue - 1);
             RuleFor(dto => dto.StockProductId).NotEmpty().InclusiveBetween(1, long.MaxValue - 1);
             RuleFor(dto => dto.SaleDescription).MaximumLength(256);
         }
