+++ MissTortasRework/MissTortas.Services/DTO/Products/SaleProductCreateDTO.cs
@@ -2,8 +2,8 @@
 {
     public class SaleProductCreateDTO
     {
-        public float SalePrice { get; set; }
-        public float SaleQuantity { get; set; }
+        public double SalePrice { get; set; }
+        public double SaleQuantity { get; set; }
         public long ProductId { get; set; }
         public string SaleDescription { get; set; } = string.Empty;
         public string SaleImagePath { get; set; } = string.Empty;
